Splitting validators unix functions into own file
